Use relative directory to simulation.py when working with templates
In out environment we have consumed Ansible roles under the "/roles" directory
in our project root. With the hardcoded playbook_dir path we can not move this
role as Git submodule to the above mentioned folder.

This commit defines a relative path to simulation.py and then joins new paths
as Python f'strings to find the templates used for rendering.